Added update site project build with Tycho
The update site uses the eclipse-repository packaging type to build a p2 repository that includes all features in the current build (maven reactor). In a second step, the mirror goal of the tycho-p2-extras-plugin [1], which is available in Tycho 0.13.0, is used to merge the existing update site into the generated p2 repository.

Due to a bug [2] in Tycho 0.13.0 which fails to add optional dependencies to the class path when compiling, the entry 'jars.extra.classpath' was added to all build.properties files of plugins which use optional dependencies (for the their tests).

[1] http://wiki.eclipse.org/Tycho/Additional_Tools
[2] https://bugs.eclipse.org/bugs/show_bug.cgi?id=351842