show refactoring menu on graph nodes
The JDT refactoring menu is added to the graph node context menu, and
the graph or package nodes are adapted to Java elements to actually
enable the refactoring actions.

There is a small, but acceptable issue: The refactoring menu shows its
normal keyboard shortcuts, which don't work in the graph view.